backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
anthropic==0.39.0
PyMuPDF==1.24.0
python-docx==1.1.0
pillow==10.4.0
pytesseract==0.3.13
httpx==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.0
